refactor(flags): merge files
