+++ internal/config/config_test.go
@@ -17,6 +17,8 @@ limitations under the License.
 package config
 
 import (
+	"bytes"
+	"errors"
 	"io"
 	"strings"
 	"testing"
@@ -333,5 +335,34 @@ func TestConfigValidate(t *testing.T) {
 		err := cfg.Validate()
 		assert.Error(t, err)
 		assert.EqualError(t, err, "invalid probe listen address: lookup tcp/invalid: unknown port")
+	})
+}
+
+func TestIsHelpRequested(t *testing.T) {
+	t.Parallel()
+
+	t.Run("returns true and writes message for HelpRequested error", func(t *testing.T) {
+		t.Parallel()
+
+		buf := &bytes.Buffer{}
+		helpMsg := "this is the help message\n"
+		err := &HelpRequested{Message: helpMsg}
+
+		ok := IsHelpRequested(err, buf)
+
+		assert.True(t, ok)
+		assert.Equal(t, helpMsg, buf.String())
+	})
+
+	t.Run("returns false and writes nothing for unrelated error", func(t *testing.T) {
+		t.Parallel()
+
+		buf := &bytes.Buffer{}
+		err := errors.New("some other error")
+
+		ok := IsHelpRequested(err, buf)
+
+		assert.False(t, ok)
+		assert.Equal(t, "", buf.String())
 	})
 }
