--- internal/config/utils.go
@@ -1,17 +1,0 @@
-package config
-
-import (
-	"errors"
-	"fmt"
-	"io"
-)
-
-// IsHelpRequested checks if the error is a HelpRequested sentinel and prints it.
-func IsHelpRequested(err error, w io.Writer) bool {
-	var helpErr *HelpRequested
-	if errors.As(err, &helpErr) {
-		fmt.Fprint(w, helpErr.Error()) // nolint:errcheck
-		return true
-	}
-	return false
-}
